third_party/WebKit/LayoutTests/fast/dom/shadow/shadow-element-inactive.html

<!DOCTYPE html>
<body>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
description('We should not expose UA shadow nodes.');

var elementNames = ['details', 'input', 'summary'];
var root2;
elementNames.forEach(function(elementName) {
    debug(elementName);
    var target = document.createElement(elementName);
    document.body.appendChild(target);
    var root1 = target.createShadowRoot();
    root1.innerHTML = '<div><shadow></shadow></div>';
    root2 = root1.firstChild.createShadowRoot();
    root2.innerHTML = '<content select="*"></content>';
    shouldBe('root2.firstChild.getDistributedNodes().length', '0');
    document.body.removeChild(target);
});
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
